android: re-enable the last disabled StartupLoadingMetricsTest
This is more relevant now because we are introducing a new flavour of
startup metrics on Android (crbug.com/1273097).

This test (#testFromExternalAppRecorded) is the most important of all,
as it is the only one checking the cases when the ".Tabbed" histograms
are recorded.

The identified reason for flake in this test is that recording the
histogram TimeToFirstNavigationCommit.Tabbed can be omitted due to a
race (if navigation happens fast). There is a separate histogram to
indicate which way the race went
(FirstNavigationCommitOccurredPreForeground). Allow both outcomes of the
race to pass and verify with the indicator histogram. The
"TimeToFirstContentfulPaint.Tabbed" has the same property, will add
indicator checks as a followup.

Bug: 1023433
Change-Id: Id8131f4be67f4321ff3fabc742cf1defc8ce4c7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4129637
Commit-Queue: Egor Pasko <pasko@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120648}