Fix vertical centering of text in holding space.
For in-progress items in holding space, a primary line of text and
a secondary line of text may be shown. With jelly enabled, they do
not have the same line height which results in incorrect vertical
centering.

This CL should have no effect when jelly is disabled since in that
case default line height is used for both labels.

Device:
* Jelly (before):    http://shortn/_fXutpPxjnZ
* Jelly (after):     http://shortn/_U67hao66aK
* No jelly (before): http://shortn/_zggYvW4IxX
* No jelly (after):  http://shortn/_zChjQqM7ar

Glinux:
* Jelly (before):    http://shortn/_XObk456dzL
* Jelly (after):     http://shortn/_fPf3e315yx
* No jelly (before): http://shortn/_arIgav6fNh
* No jelly (after:   http://shortn/_duATt7k2lA

Bug: b:273982593
Change-Id: I577b7a419e576b59775d08699ac668359319e745
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4348198
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: David Black <dmblack@google.com>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120680}