Add browser finder helper function to find all of a profile's browsers
This adds a helper function to find all of the browser windows for a
profile. It is very similar to the existing function
FindAllTabbedBrowsersWithProfile(), except it matches all browsers,
not just tabbed/normal browsers.

This is needed for the Download Bubble.

Change-Id: I2c81dc10a7ef34af7e6962c0bb49ca026a413218
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4362140
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120681}