Reland "[fyfre]Block browser launch when first run is running on Mac"
This is a reland of commit 267825666fecc24447a7cf8f41f3a71a76553cfb

The failure was that we were calling LaunchBrowserStartup instead of CreateBrowser in applicationShouldHandleReopen. This wasn't needed because we will never reach this call and it was causing failures in the tests because LaunchBrowserStartup and CreateBrowser currently have different behaviors.

Original change's description:
> [fyfre]Block browser launch when first run is running on Mac
>
> Fixed: 1423299
> Change-Id: Iadfbdbed6b49d664a98e275cb17fd81849a3aa73
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4335540
> Reviewed-by: David Roger <droger@chromium.org>
> Commit-Queue: Jack Yammine <jyammine@google.com>
> Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1120455}

Change-Id: Id5dd0b14443791616b9104441029360fa9ec12d0
Fixed: 1423299
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4361247
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Jack Yammine <jyammine@google.com>
Cr-Commit-Position: refs/heads/main@{#1120694}