Add llvm-profile-file switch
This switch will be used on Lacros PGO builder to train Lacros
PGO profiles.
Lacros is similar to Android that it's running on a device that is not
the host machine and environment variables aren't well supported.
But Lacros also need to pass in the path so it is the same path as
isolate test output folder on bots.

Bug: 1417071
Change-Id: If68743a70e94f108eb6f6ed23f3aa99d3c2aebf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4358910
Commit-Queue: Sven Zheng <svenzheng@chromium.org>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120714}