Fix bug where a prefetch was marked as timed out after being served.
When a prefetch takes longer than a given timeout duration (default of
10 seconds), it fails and is marked with a completion status of
ERR_TIMED_OUT. This included if the prefetch was served and the response
was being streamed, which is not the desired behavior. This CL fixes
that bug, by stopping the timeout timer once a prefetch is served.

Bug: 1426632
Change-Id: I313137c484f8feb72d0b981c018a016f6ca38cf3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4358563
Commit-Queue: Liviu Tinta <liviutinta@chromium.org>
Auto-Submit: Max Curran <curranmax@chromium.org>
Reviewed-by: Liviu Tinta <liviutinta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120724}