[Reland][Extensions] Don't allow unregistration of extension service workers
This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/4257067.

--- Reland Notes ---
The original version caused the browser test
ServiceWorkerTestWithEarlyReadyMesssage.MissingRegistrationMitigated to
be flaky.

This wasn't entirely unexpected, since that CL relies on removing the
SW registration, which this CL (intentionally) makes more difficult. In
testing, the previous approach (immediately trying to remove the
registration) consistently worked, but it seems it was not quite
robust enough.

Instead, add a test method to temporarily allow unregistering certain
extension service workers.

--- Original Description ---

If an extension is service worker-based, the registration of its
root-scope service worker (the extension's primary service worker) is
critical to the extension's functionality; removing it results in the
extension being in a broken state. Given this, we should not allow the
unregistration of this service worker through "normal" means (such as
the browsingData API). The extension service worker's registration
should instead be tied to the extension's enabled state (registered
as long as the extension is enabled). If a user or other party wants
to unregister this worker, they should disable or remove the extension.

This CL introduces a new method on ContentBrowserClient,
`MayDeleteServiceWorkerRegistration()`, which allows embedders to
indicate that a service worker registration cannot be deleted. Check
this prior to deleting a service worker, and disallow deletion of any
root-scope service workers associated with enabld service worker-based
extensions.

Add tests to cover both an example case that should now be impossible
(deleting an extension service worker through the browsingData API), as
well as other scenarios that should keep working (like updating and
removing the extension) to ensure these flows are properly handled (and
do allow deletion of the service worker).

Known limitation: the button for "Unregister" in
chrome://serviceworker-internals is still visible, but clicking it
does nothing (but also, doesn't crash). A followup will remove the
button from the UI.

Bug: 1392498
Change-Id: Ife518812a8d9e2847d51a267724b2a0f7674e337
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4354471
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120732}