[Test Automation] Make AddStep(MultiStep&, ...) available in test code
These are useful utilities that allow addition of steps to existing step sequences. It was previously private because it was unclear if it would ever be needed in test code, but we believe it will be useful when procedurally generating test sequences.

This change allows for test code to do the following:

  MultiStep steps = Steps(...);
  if (some_condition) {
    AddStep(steps, ClickButton(kButtonId));
  }
  return steps;

Without it, you would have to write the following, which is less clear:

  MultiStep steps = Steps(...);
  if (some_condition) {
    steps = Steps(steps, ClickButton(kButtonId));
  }
  return steps;

We use this kind of logic all the time internally. However we've realized it might be useful in e.g. a fuzzing application.

Change-Id: I7a0d92b74b1ca62db3a69f9ede1ea61af4d07fa0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4358558
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Brendon Tiszka <tiszka@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120734}