Add support for OverlayDelegates to specify a desired rotation
Exposes the ability for XrImmersiveOverlay.Delegate to express a desired
rotation and modifies the XrImmersiveOverlay to handle and determine if
a rotation is necessary. For Ar we simply just want to lock the existing
rotation; but for VR, we need to rotate to landscape which may require
a rotation *actually* being triggered and also requires us to transpose
the width/height since we don't get any further update when the rotation
actually occurs.

Bug: 989117
Change-Id: I189a0683181090a083703029622cd7c97c2a5ad3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4358554
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Commit-Queue: Piotr Bialecki <bialpio@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120752}