[Adaptive Toolbar] Hook-up "Add to bookmarks" button to UI and settings
This CL adds an "Add to bookmarks" button to the adaptive toolbar
page in settings. This setting will be hidden behind a flag, so users will need to enable it in chrome:flags before they can select it in settings.

This CL also adds an IPH message to the button.

Bug: 1425502
Change-Id: I2fbd7c94c3d5ec0ff3ec672ff9573e593bf10112
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4348520
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Commit-Queue: Salvador Guerrero Ramos <salg@google.com>
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120799}