Ignore time travel when simulating_outdated_
This ignores build-time overrides that may potentially be in the future.
It's very possible that we still hit DumpWithoutCrashing() even with
this guard present (because local time may be far in the past), but this
should drastically reduce the number of dumps so that we can look at the
remaining uploads.

Bug: 1407664
Change-Id: I7d1a55c228be12dfdcb264016b10303e200aa512
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4363338
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Peter Boström <pbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120816}