[FLEDGE] Support Data-Version in bidding signals when there are no keys.
When there's a trustedBiddingSignalsUrl but trustedBiddingSignalsKeys
is empty, we now fetch trustedBiddingSignalsUrl for other uses. In
this case, however, we were throwing away the Data-Version header, if
present. This CL makes us pass the Data-Version to the bidder worklet
in that case (which also better matches the wording of the spec).

Also reject trusted signals responses that are JSON arrays, since
they're expected to be JSON objects instead. Previously, they were
allowed because V8's parser, and Javascript more generally, consider
Arrays to be a subtype of Object.

Bug: 1426832
Change-Id: Ie84ead9f24436e95026ea5df2635b0cc3f1f3c2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4359778
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Russ Hamilton <behamilton@google.com>
Cr-Commit-Position: refs/heads/main@{#1120844}