Add a11y description to announce the number of apps in a folder
In the folder icon UI refresh, we added the app counter on the folder
icon so that the users can see how many apps are in a folder at the
first glance. To provide an equitable experience for chromevox users,
This cl adds the a11y string to announce the number of apps in a folder
when focusing on it.

screenshot: https://screenshot.googleplex.com/64fUyz6n8gnWPwj

Bug: b:274728929
Change-Id: If9d7c03901673291cf204f2dad12d840e8b8ddbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4362800
Commit-Queue: Wen-Chien Wang <wcwang@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120847}