Remove base_unittests the on dev android builder
And replace it with chrome_public_smoke_test. base_unittests doesn't
appear to run on any sheriff'ed arm-based android device bot. This led
to the suite failing for days due to a regression. (It was finally
fixed as part of crbug.com/1423829, but as part of rust spin-up.)

So this replaces it with a smoke suite that's hopefully more
sheriff'ed.

This also renames the test suite groupings from 'swarm' to 'dev'.

Bug: 1426952
Change-Id: I76eade6319d780579c377b46de236a7261ca03ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4363602
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Haiyang Pan <hypan@google.com>
Reviewed-by: Haiyang Pan <hypan@google.com>
Cr-Commit-Position: refs/heads/main@{#1120854}