[Share] Fix detaching too early when ShareActivity is finished
TargetChosenReceiver#onIntentCompleted is called immediately when Android share sheet activity finished, which can be earlier than the pending intent for the chooserIntent is sent, where #onReceive is called properly. In those cases, the callback is never correctly invoked before the broadcast receiver is detached.

This CL fix this issue by only detach the broadcast receiver when the share is canceled. If share is not canceled, the broadcast receive will stick around until the pending intent is received, or the next share sheet is triggered.

Note that there's one edge case, when "Copy" is selected from OS share sheet, the receiver is not detached correctly and it might live longer than necessary in memory. This does not impact the functionality of ShareHelepr, since the last receiver will be detached before the next share sheet is shown in #sendChooserIntent. As reference, it was the status-quo before the change in crrev.com/c/4338839.

Bug: 1421447, 1410201
Change-Id: I5d5818b1de10c4829d7aa36afa7c49c04f1215a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4362999
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120863}