Soften some language in styleguide/c++/checks.md
A paragraph saying that we reserve the right to upgrade DCHECKs to
CHECKs is removed as it suggests that changes will be made to DCHECKs
under the hood bypassing code reviewers. This may have been discouraging
the addition of DCHECKs in the first place.

The intent here is to suggest CHECKs over DCHECKs, it should never be
interpreted as discouraging the addition of CHECKs in the first place.

Bug: None
Change-Id: Ic81114de969e6a9c33c0f4e43a01d3aa49a1f3b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4363704
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Boström <pbos@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120887}