shortcuts: Update SearchBox and SearchResultRow styles and behavior to match spec
This CL updates the general styles and focus styles of the SearchBox and
SearchResultRow components to match the spec and the Settings app.

The visual changes can be seen in the "After" screenshots below.
Much of the CSS is adapted from the Settings code to maintain a
consistent style.

The main behavior change is that clicking outside of the search box
now closes the dropdown. I also added a test for this behavioral change.

Before:
https://screenshot.googleplex.com/9ZpCywbguTUVrz8
https://screenshot.googleplex.com/BpjcUoBkVDdTEa2
https://screenshot.googleplex.com/3dRMPYcqAAW9WFN

After:
https://screenshot.googleplex.com/6L3eNLd5x7xJQGx
https://screenshot.googleplex.com/9nJt4DErtQ5NrQs
https://screenshot.googleplex.com/4NWQ9vtgGt7vxTX
https://screenshot.googleplex.com/5eEZroUKqGsCo5V

Bug: b:266085562
Change-Id: I1f294a5b88101b52527d17612eddf19d26c70b8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4327408
Commit-Queue: Camden Bickel <cambickel@google.com>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120934}