Disable optional toolbar button animations when not changing variants
The new bookmark toolbar button changes its icon depending on whether
the current page is bookmarked or not. Right now this icon update is
animated, which can be distracting and unexpected.

This CL modifies OptionalButtonView to block animations when updating
within the same button variant (e.g. New Tab, Bookmark, Share).

An additional effect of this change is that the profile disc won't
animate when updating from the profile picture placeholder to the
real user picture.

Bug: 1425502
Change-Id: I69f0a081b3575e55fef6a2543108a0871b5bd404
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4361940
Commit-Queue: Salvador Guerrero Ramos <salg@google.com>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1120941}