fileSystemProvider: pass provider icons over crosapi
chrome://extension-icon/ and chrome://app-icon/ resources for Lacros
extensions are not available in Ash, an extension implementing
fileSystemProvider in Lacros needs to pass the icon to Ash some other
way.

Add a new version of the ExtensionLoaded crosapi to pass provider's
icons from Lacros to Ash. The icons are encoded as PNG data URLs.

Bug: b:266519645
Change-Id: I7f55023642fc4bb4a479b91ce4268e1e7135916f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4349298
Commit-Queue: Alexander Bolodurin <alexbn@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1121022}