[MediaPicker]: Add four variations for launch.
It is not entirely clear which way of launching the Media
Picker is the right one for Chrome, so we're adding four
variations to allow for some experimentation.

The four variations are:

1) Launch indirectly (via ACTION_GET_CONTENT).
2) Launch directly (via ACTION_PICK_IMAGES).
3) Same as 2, but supplies intent w/complex MIME types.
4) Stick to Chrome Picker, but without the Browse button.

The fourth option will only be used as a last resort (if
the changes to ACTION_GET_CONTENT affect our Browse in such
a way that we'd rather remove the Browse button than suffer
the UX changes).

This CL also deprecates the old feature flag.

Bug: 1342029
Low-Coverage-Reason: Three variations modify calls for external intents, which launch the Media Picker (an external app outside our testing scope). That leaves one variation, which makes changes to Chrome (and is covered by a test).
Change-Id: I3580bc1d960af30a5cf53c40fcf41e25cbc2e000
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4360520
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Rayan Kanso <rayankans@chromium.org>
Auto-Submit: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1121052}