[BSC]: Add on demand cookie refresh functionality
This CL adds on demand cookie refresh to the API of
`BoundSessionCookieRefreshService`. This is an asynchronous operation.
The callback is called synchronously if there is no bound session.
Otherwise, the refresh is delegated to the
`BoundSessionCookieController`.

- If there is a fresh cookie, the callback is called synchronously.
- If there is a pending refresh, the callback will be appended to the
  list of callbacks requesting a fresh cookie.
- If no pending requests, it starts a new refresh request.
- All pending callbacks are called once the cookie is known to be fresh,
  the request succeeds/fails or the session is terminated.

This feature will be used by network request throttlers to request a
cookie refresh, the callback resumes the deferred requests.

Follow-up: Add timeout for unblocking requests/Running all callbacks.

Bug: b/263264391, 1392135
Change-Id: Ie1832d0f0281ca9ce1b0dc3929b8fa7cf9adb9e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4329788
Commit-Queue: Monica Basta <msalama@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1121089}