[Pepper] Document ForceEnablePepperVideoDecoderDevAPI correct default
The ForceEnablePepperVideoDecoderDevAPI landed with a documented default
value of true, whereas the actual default value in the code is false
(i.e., the default value of the pref controlling the policy is false).
This CL changes the documented default value to be false to match the
code.

Note: This documented default value does not show up in the client-side
documentation of the policy. It may be used internally, and I am
following up on whether there are any internal toggles that need to be
changed if so.

Bug: 1427069
Change-Id: I5b61fabf2671c7b215c9a5d61015021dcc27bc2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4364980
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1121091}