[sync] DualLayerUserPrefStore: Notify only if effective value changes
This CL wraps up a few-todos to notify the observers only when the
effective value of a pref changes.

Without this check in RemoveValue(), the usage of DualLayerUserPrefStore
can lead to a recursive callback and thus, a crash. This CL,
consequently, unblocks crrev.com/c/4358058.

Bug: 1416477
Change-Id: I9ad196764ca1534f7cfd54092a46585f73d74982
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4364838
Commit-Queue: Ankush Singh <ankushkush@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1121094}