[TrustedVault] Fix SetPrimaryAccount() logic during startup
This is a fix for crrev.com/c/4280636, that intended to call
SetPrimaryAccount() regardless of whether refresh tokens are loaded.
Without calling this in IdentityManagerObserver ctor.
There are two fixes:
1. SetPrimaryAccount() wasn't called on startup, if primary account was
set before TrustedVaultClient is created and refresh tokens are not yet
loaded. Now it's called IdentityManagerObserver ctor.
2. Upon refresh tokens loading, the backend might be not notified, if
primary account was already set. This is because OnRefreshTokensLoaded()
calls UpdatePrimaryAccountIfNeeded() and the latter optimized the call,
since primary account wasn't changed. Now it calls
OnErrorStateOfRefreshTokenUpdatedForAccount().

Bug: 1413179
Change-Id: I1646209efb9f3468c77bdec1121113a209312892
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4353057
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Commit-Queue: Maksim Moskvitin <mmoskvitin@google.com>
Cr-Commit-Position: refs/heads/main@{#1121103}