[Reland]Fix the mic mute toast with the revamped view
This is the reland cl of
https://chromium-review.googlesource.com/c/chromium/src/+/4357636 .

I couldn't reproduce the asan test failure locally. But I think
I've found the root cause of it: the raw pointer didn't get destroyed
manually after it's created.

I have filed a bug: b/274820054, which will fix the base class
function to return an unique pointer instead of a raw pointer.

******************************************************
Original description:

The mute button was replaced by the icon container in the revamped
slider view. However, we still need to show this button in the mute
toast view.

In this cl we add the button and hide the container when it's the
toast bubble in the revamped view.

screenshots:
https://screenshot.googleplex.com/BLCCrjC99t7NDDz

Bug: b/273986448
Change-Id: I487443e0bb2896cc6586d399a9b6101175ba9561
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4363701
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Alex Newcomer <newcomer@chromium.org>
Auto-Submit: Jiaming Cheng <jiamingc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1121134}